Fix for x86_64.
	* src/engine/qofquerycore.c: Fix for x86_64.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@10281 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/qofquerycore.c
@@ -859,7 +859,7 @@ int64_to_string (gpointer object, QofParam *getter)
 {
   gint64 num = ((query_int64_getter)getter->param_getfcn)(object, getter);
 
-  return g_strdup_printf (GNC_SCANF_LLD, num);
+  return g_strdup_printf (GNC_SCANF_LLD, (long long int) num);
 }
 
 /* ================================================================ */
